+++ module-system/module-implementations/integration-tests/src/attester_incentives/process_attestation.rs
@@ -9,7 +9,7 @@ use sov_modules_stf_blueprint::TxEffect;
 use sov_state::{DefaultStorageSpec, StorageRoot};
 use sov_test_utils::attester_incentive_data::AttesterIncentivesMessageGenerator;
 use sov_test_utils::runtime::TestRuntime;
-use sov_test_utils::{new_test_blob_from_batch, MessageGenerator};
+use sov_test_utils::{new_test_blob_from_batch, MessageGenerator, TestHasher};
 
 use super::{AttesterIncentivesTestHandler, USER_BALANCE};
 use crate::helpers::{Da, ExecutionSimulationVars, TestRollup, S};
@@ -59,7 +59,7 @@ impl AttesterIncentivesTestHandler {
     // Attest only the first transition and check that the attestation was processed correctly
     fn try_attest_first_transition(
         &self,
-        genesis_root: StorageRoot<DefaultStorageSpec>,
+        genesis_root: StorageRoot<DefaultStorageSpec<TestHasher>>,
         execution_vars: Vec<ExecutionSimulationVars>,
         honest_attester_balance: u64,
         rollup: &mut TestRollup,